Revert "Add Value op, fix, refine"
This reverts commit a19f0760a986b74b998bb135622d7664bc9bb7ff.

+++ include/mukyu/cablin/core/config.hpp
@@ -96,7 +96,7 @@ inline std::string Config::as<std::string>() const {
 
 template <>
 inline float Config::as<float>() const {
-    return asFloat();
+    return asInt();
 }
 
 template <>
